Create mobile_app_icon_fixer.sh
The purpose of this script is to find any mobile device applications in the JSS that do not
have an icon associated with them, and upload an icon for that application. As with any
script that makes bulk changes to the JSS, it is advised to make a database backup.