Update to bypass web certificate SSL validation
There were issues when the SSL certificate was not trusted from the
client running the script.  This update removes SSL validation for the
JSS.  It also appears that some unknown character returns in the script
that were removed.  This is why the whole script looks new.  The code
change on the SSL side is the addition of the line:
gcontext = ssl._create_unverified_context()

and including the gcontext to all HTTP requests as such:
urlopen(req, context=gcontext)